refactor inbox
